Explicitly define IP as an 8-bit pointer
Atmel Studio had the "-fshort-enums" option on, whereas the Arduino
compiler was using the standard (int-sized) enums.  This meant that the
instruction pointer was 16-bits on the Arduino and so ip++ was skipping
over two bytes instead of one byte.

The instruction pointer is now an 8-bit value at all times, ensuring
consistency between the two environments (and correct operation on
Arduino).