True indirect threading
Previously the CFA contained a jump table index that pointed at the
address of the DO* label that implemented the CFA.  Now the CFA is a
two-byte value that contains the address of the label; a true direct
threaded interpreter.

This massively speeds up the address interpreter, because we can just
jump to the CFA instead of first having to look up the address in a
table.

The only downside is that we cannot get the address of the label outside
of the innerInterpreter function.  That won't be a problem when
arfOpCOLON has been defined, but at the moment it means that we have to
manually find and copy the address of DOCOLON into our test dictionary
after every change to the file.