Non-curses test tool for use in Cygwin
We actually can use (n)curses in Cygwin, but then we run into problems
with GDB -- GDB apparently doesn't end lines with \r\n and once we go
into curses mode Cygwin needs that \r in order to display things
properly.  There is probably a way to put curses in the right mode, but
for now I'll just use this line-oriented test tool.