Eliminate CFAs; DO* are now standard primitives
Minimizes register usage and allows us to have a single jump table.
Doubles the number of available tokens.  Simplifies the inner
interpreter and paves the way for ROM-based definitions (instead of the
current copy-to-RAM definitions).

Makes call sites one byte longer, although this can be mitigated per
notes in the code.