Start moving primitive defs into a single block
Among other things, this allows us to use a single C++ code block to
handle all of these definitions instead of duplicating the code
everywhere.  We'll probably want to auto-generate the token order so
that these can be packed in as tightly as possible, but for now we'll
just put them at the end in order to keep making progress.