Fix UD/MOD on Arduino (for most inputs...)
The previous code didn't work at all, whereas the new code (which is
really just the old code) works most of the time.  The new (old) code
only appears to fail in extreme cases such as ((65535 * 65535) / 65535).
My guess is that this is a compiler difference between whatever gcc is
used by Atmel Studio 6.2 and the gcc used by Arduino 1.0.5.  Maybe
Arduino 1.5 is better?