ROM Definitions now use the new dictionary format
LFAs are now XTs, names are in reverse order (just like MFORTH), and the
inner interpreter can go from an XT to the CFA/PFA with basic addition.
User definitions are still using the old format and so they cannot be
called from other definitions (although it happens that they can still
be EXECUTEd).  All of the tests are broken.